+++ 21H2/Wdf01000/-HandleQueryPnpDeviceState@FxPkgPnp@@IEAAKK@Z.c
@@ -1,13 +1,13 @@
 /*
- * XREFs of ?HandleQueryPnpDeviceState@FxPkgPnp@@IEAAKK@Z @ 0x1C0010B10
+ * XREFs of ?HandleQueryPnpDeviceState@FxPkgPnp@@IEAAKK@Z @ 0x1C0080BD0
  * Callers:
- *     ?HandleQueryPnpDeviceStateCompletion@FxPkgFdo@@AEAAXPEAVFxIrp@@@Z @ 0x1C00109B0 (-HandleQueryPnpDeviceStateCompletion@FxPkgFdo@@AEAAXPEAVFxIrp@@@Z.c)
- *     ?_PnpQueryPnpDeviceState@FxPkgPdo@@CAJPEAVFxPkgPnp@@PEAVFxIrp@@@Z @ 0x1C0032010 (-_PnpQueryPnpDeviceState@FxPkgPdo@@CAJPEAVFxPkgPnp@@PEAVFxIrp@@@Z.c)
+ *     ?_PnpQueryPnpDeviceState@FxPkgPdo@@CAJPEAVFxPkgPnp@@PEAVFxIrp@@@Z @ 0x1C0079B10 (-_PnpQueryPnpDeviceState@FxPkgPdo@@CAJPEAVFxPkgPnp@@PEAVFxIrp@@@Z.c)
+ *     ?HandleQueryPnpDeviceStateCompletion@FxPkgFdo@@AEAAXPEAVFxIrp@@@Z @ 0x1C008427C (-HandleQueryPnpDeviceStateCompletion@FxPkgFdo@@AEAAXPEAVFxIrp@@@Z.c)
  * Callees:
- *     ?IsInSpecialUse@FxPkgPnp@@IEAAEXZ @ 0x1C0010C50 (-IsInSpecialUse@FxPkgPnp@@IEAAEXZ.c)
- *     _guard_dispatch_icall_nop @ 0x1C0036BA0 (_guard_dispatch_icall_nop.c)
- *     ?DecrementPendingTaskCount@FxCompanionTarget@@AEAAXXZ @ 0x1C0060FD4 (-DecrementPendingTaskCount@FxCompanionTarget@@AEAAXXZ.c)
- *     ?IncrementPendingTaskCountWithFloor@FxCompanionTarget@@AEAAJXZ @ 0x1C0061190 (-IncrementPendingTaskCountWithFloor@FxCompanionTarget@@AEAAJXZ.c)
+ *     _guard_dispatch_icall_nop @ 0x1C001D510 (_guard_dispatch_icall_nop.c)
+ *     ?DecrementPendingTaskCount@FxCompanionTarget@@AEAAXXZ @ 0x1C0043604 (-DecrementPendingTaskCount@FxCompanionTarget@@AEAAXXZ.c)
+ *     ?IncrementPendingTaskCountWithFloor@FxCompanionTarget@@AEAAJXZ @ 0x1C00437C4 (-IncrementPendingTaskCountWithFloor@FxCompanionTarget@@AEAAJXZ.c)
+ *     ?IsInSpecialUse@FxPkgPnp@@IEAAEXZ @ 0x1C0079EE4 (-IsInSpecialUse@FxPkgPnp@@IEAAEXZ.c)
  */
 
 __int64 __fastcall FxPkgPnp::HandleQueryPnpDeviceState(FxPkgPnp *this, unsigned int PnpDeviceState)
@@ -24,7 +24,7 @@ __int64 __fastcall FxPkgPnp::HandleQueryPnpDeviceState(FxPkgPnp *this, unsigned 
   if ( this->m_InternalFailure || this->m_Failed )
     v5 |= 4u;
   m_CompanionTarget = this->m_CompanionTarget;
-  if ( m_CompanionTarget && FxCompanionTarget::IncrementPendingTaskCountWithFloor(m_CompanionTarget) >= 0 )
+  if ( m_CompanionTarget && (int)FxCompanionTarget::IncrementPendingTaskCountWithFloor(m_CompanionTarget) >= 0 )
   {
     m_CompanionTarget->m_RdDeviceCompanion->QueryPnPDeviceStateNotification(m_CompanionTarget->m_RdDeviceCompanion);
     FxCompanionTarget::DecrementPendingTaskCount(m_CompanionTarget);
